+++ Changelog.md
@@ -8,6 +8,7 @@ Rust:
     - Add convert gate
 - remove ExpandDefinable
 - remove gate_set and features from Relation struct
+- remove Boolean gates (and, xor, not)
 
 # SIEVE IR v1.0.1, 2022-08